requirements.txt
fastapi==0.115.12
gradio==4.44.1
gradio_client==1.3.0
locust
motor==3.7.0
numpy
pandas==2.0.0
pydantic-settings==2.9.1
pymongo==4.12.0
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.31.0
tiktoken==0.9.0
PyJWT==2.10.1
pytesseract
discord
diffusers
torch
transformers
accelerate 
pillow
gradio
torch
transformers
diffusers
Pillow
accelerate 
peft
torchao
bitsandbytes
huggingface_hub
python-dotenv
sentencepiece 
protobuf


